added Announcement repository
